fix(ai): AI/画像解析の堅牢性を強化 (#1047)
- fast-llm(Grok) 呼び出しの signal 誤用を修正しタイムアウトを付与
  (messages/route.ts の signal ボディ混入、blood-tests/checkups 未設定を解消)
- 重要度チェックの fire-and-forget を応答前 await(5秒上限)に変更し
  serverless での消失を防止
- アクション自動実行の self-fetch を廃止し、実行ロジックを
  consultation-action-executor.ts に抽出して直接呼び出す
  (APP_URL 未設定時の静かな失敗・pending 滞留を解消)
- add_recipe_to_collection のクエリビルダーを列値として渡す死にコードを削除
- gemini-json.ts (edge/Next 両方) に実行時スキーマ検証を追加し、
  型不一致・enum外の値を検出してリトライへフォールスルーさせる
- evidence-verifier.ts で参照値との乖離50%超(excessive_deviation)の場合に
  recommendedCalories へクランプし、nutrition-pipeline.ts で栄養素合計を補正
