+++ src/app/api/health/checkups/route.ts
@@ -175,6 +175,9 @@ eGFR: ${checkup.egfr ?? '-'} mL/min/1.73m²
     ],
     response_format: { type: 'json_object' },
     max_tokens: 1000,
+  }, {
+    // #1047 F2-15: fast-llm 呼び出しにタイムアウトが無かった
+    signal: AbortSignal.timeout(25_000),
   });
 
   const content = response.choices[0]?.message?.content;
@@ -259,6 +262,9 @@ LDL: ${c.ldl_cholesterol ?? '-'} mg/dL
     ],
     response_format: { type: 'json_object' },
     max_tokens: 2000,
+  }, {
+    // #1047 F2-15: fast-llm 呼び出しにタイムアウトが無かった
+    signal: AbortSignal.timeout(25_000),
   });
 
   const content = response.choices[0]?.message?.content;
