+++ supabase/functions/_shared/evidence-verifier.ts
@@ -6,7 +6,7 @@
  */
 
 import { SupabaseClient } from "@supabase/supabase-js";
-import { NutritionTotals } from './nutrition-calculator-v2.ts'
+import { NutritionTotals, roundNutrition } from './nutrition-calculator-v2.ts'
 import { isRetryableError, withRetry, withTimeout } from './network-retry.ts'
 
 // ============================================
@@ -36,6 +36,15 @@ export interface VerificationResult {
   warning: string | null
   referenceRecipe: ReferenceRecipe | null
   allReferences: ReferenceRecipe[]
+  /**
+   * #1047 F5: 保存に使用すべきカロリー値。
+   * reason==='excessive_deviation'（参照値との乖離50%超）の場合のみ、
+   * calculatedCalories を参照値の ±MAX_DEVIATION_RATIO の範囲にクランプした値。
+   * それ以外は calculatedCalories と同じ。
+   * 以前は乖離50%超でも calculatedCalories（未補正の生値）がそのまま保存され、
+   * confidenceScore を下げるだけで実際の保存値は補正されていなかった。
+   */
+  recommendedCalories: number
 }
 
 export interface EvidenceInfo {
@@ -122,6 +131,18 @@ function calculateDeviation(calculated: number, reference: number): number {
   return Math.abs((calculated - reference) / reference) * 100
 }
 
+// #1047 F5: 参照値との乖離が50%を超える場合にクランプする許容比率。
+// 20-50%（high_deviation）は警告付きでそのまま採用するが、50%超（excessive_deviation）は
+// 参照値の ±50% の範囲に丸め込む（完全に参照値へ置き換えると計算結果を無視することになるため、
+// 「明らかに信頼できる範囲」まで寄せるに留める）。
+const EXCESSIVE_DEVIATION_CLAMP_RATIO = 0.5
+
+function clampToReferenceRange(calculated: number, reference: number): number {
+  const min = reference * (1 - EXCESSIVE_DEVIATION_CLAMP_RATIO)
+  const max = reference * (1 + EXCESSIVE_DEVIATION_CLAMP_RATIO)
+  return Math.min(Math.max(calculated, min), max)
+}
+
 // ============================================
 // 検証実行
 // ============================================
@@ -137,6 +158,7 @@ export async function verifyNutrition(
     return {
       isVerified: false,
       calculatedCalories: calculatedNutrition.calories_kcal,
+      recommendedCalories: calculatedNutrition.calories_kcal,
       referenceCalories: null,
       deviationPercent: null,
       reason: 'no_reference',
@@ -153,6 +175,7 @@ export async function verifyNutrition(
     return {
       isVerified: false,
       calculatedCalories: calculatedNutrition.calories_kcal,
+      recommendedCalories: calculatedNutrition.calories_kcal,
       referenceCalories: null,
       deviationPercent: null,
       reason: 'no_reference',
@@ -169,6 +192,7 @@ export async function verifyNutrition(
     return {
       isVerified: true,
       calculatedCalories: calculatedNutrition.calories_kcal,
+      recommendedCalories: calculatedNutrition.calories_kcal,
       referenceCalories: refCalories,
       deviationPercent: Math.round(deviation * 10) / 10,
       reason: 'ok',
@@ -177,10 +201,11 @@ export async function verifyNutrition(
       allReferences: references,
     }
   } else if (deviation <= 50) {
-    // 20-50%: 警告付きで採用
+    // 20-50%: 警告付きで採用（値は補正しない）
     return {
       isVerified: true,
       calculatedCalories: calculatedNutrition.calories_kcal,
+      recommendedCalories: calculatedNutrition.calories_kcal,
       referenceCalories: refCalories,
       deviationPercent: Math.round(deviation * 10) / 10,
       reason: 'high_deviation',
@@ -189,10 +214,13 @@ export async function verifyNutrition(
       allReferences: references,
     }
   } else {
-    // 50%超: 大幅な乖離
+    // #1047 F5: 50%超の大幅な乖離は confidenceScore を下げるだけでなく、
+    // 参照値の ±50% の範囲にクランプした値を recommendedCalories として提示する。
+    // 呼び出し側（nutrition-pipeline.ts）はこの値を使って保存する栄養素合計を補正する。
     return {
       isVerified: false,
       calculatedCalories: calculatedNutrition.calories_kcal,
+      recommendedCalories: clampToReferenceRange(calculatedNutrition.calories_kcal, refCalories),
       referenceCalories: refCalories,
       deviationPercent: Math.round(deviation * 10) / 10,
       reason: 'excessive_deviation',
@@ -270,4 +298,32 @@ export function createEvidenceInfo(
     verification,
     confidenceScore,
   }
+}
+
+// ============================================
+// 乖離補正の適用（#1047 F5）
+// ============================================
+
+/**
+ * verifyNutrition() の結果が excessive_deviation の場合、mealTotals 全体を
+ * recommendedCalories / calculatedCalories の比率で一律スケーリングして返す。
+ * カロリーだけを補正すると、他の栄養素（P/F/C等）との整合が崩れる
+ * （例: カロリー300kcalなのにタンパク質は600kcal相当のまま）ため、
+ * 相対的な栄養バランスを保ったまま全体を補正する。
+ * excessive_deviation でない場合は mealTotals をそのまま返す（no-op）。
+ */
+export function applyCalorieCorrection(
+  mealTotals: NutritionTotals,
+  verification: VerificationResult,
+): NutritionTotals {
+  if (verification.reason !== 'excessive_deviation') return mealTotals
+  if (verification.calculatedCalories <= 0) return mealTotals
+  if (verification.recommendedCalories === verification.calculatedCalories) return mealTotals
+
+  const ratio = verification.recommendedCalories / verification.calculatedCalories
+  const scaled = Object.fromEntries(
+    Object.entries(mealTotals).map(([key, value]) => [key, typeof value === 'number' ? value * ratio : value]),
+  ) as unknown as NutritionTotals
+
+  return roundNutrition(scaled)
 }
